This fixes #5 for keeping a running score based on the specified parameters.
It also improves on #8 for intervals. That issue still needs to be
revisited after the meeting.

Additionally, I've added a level1.xml file based on the excerpt from the
PowerPoint so it can be tested against. However, the saxophone score is
wrong. Something to discuss in the meeting.

All of this is in xwsxethan/MusicScoring