Reorganized some source files and implemented a build mechanism.
This closes #32 in xwsxethan/MusicScoring.

I'll need to come back to this to fix the running issue. Currently, to
run the generated jar, you need to execute the following command from
the XMLtoScore folder:

"java -jar ../Executables/MusicScoring.jar FileNameGoesHere.xml
DifficultyLevelFromOneToFiveGoesHere"

The reason for having to be in that specific folder is that some of the
internal file references become misaligned when the user isn't in that
code directory. That will need to be fixed later.