Setup the beginner.xml config file and tested it with the new setup.
Everything is running smoothly now. One brief bug arose with parsing the
"strong-accent" articulation due to it being named with a dash in xml
and without a dash in java, but that was since corrected. Both the old
method and the new produce the same score. The old code (for this and
the other visitor pattern change) was moved to a new folder just for
dead code, in case it may be needed to revert to at a later time.

Closes xwsxethan/MusicScoring#2