Implemented the basic web page functionality, including the following:
Navbar with links to the github page for info.
Dropdown selector to choose between two pre-defined xmls to run.
Output box to say which xml is currently selected.
Button to execute the java code for the specified file.
Output box to say what the score was once the code finishes.

Closes xwsxethan/MusicScoring#30