Added code for parsing all the parts within the score.
Only the first is output as of now, but all are available. Code to print
them all was stubbed out. Later I'll need to associate each part with
the name of an instrument/section of instruments/family of instruments
so that the score is meaningful. 

I could also start developing an overall score for the entire piece
which could be an average of the total scores or something similar.

Closes xwsxethan/MusicScoring#44