Adds tracking for worst measure for every part. Also stubs out naming for parts.
This moved the previous tracking functionality from the old status
report into a separate class. That class now owns the overall score as
well as all the smaller measurements. It has a name associated with it
that is not currently being set. Those names will have to be parsed from
the part-list and associated with the part (future work required).

Closes xwsxethan/MusicScoring#39