Cleaning up the build, run, and deployment process.
The edit removes the "The Hobbit" files from the top of git and places
them under the sample pieces folder as necessary.

This also moves the ConfigFiles directory out to WebCode (instead of
XMLtoScore). Similarly it sets up an ant auto-build on every change of
code. This build is now pushed to WebCode (instead of
XMLtoScore/Executables, which is now gone). The default MusicXML files
were also moved to WebCode/MusicXMLs/. To allow for this to work, all
future executions from the source code need to set their working
directory to WebCode and add the MusicScoring.jar file to their source
(as the primary source). Most all of this should be handled by eclipse
correctly once pushed, but if there are errors, return here for the
steps to correct them.